requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
django==5.2.1
django-debug-toolbar==4.4.6
gunicorn==21.2.0
h11==0.14.0
joblib==1.4.2s
numpy==2.1.1
packaging==23.2
pandas==2.2.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.5.0
typing-extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.0.post1
whitenoise==6.6.0